Pack dependency component
